+++ packages/stripe/lib/src/stripe.dart
@@ -141,6 +141,18 @@ class Stripe {
     return isSupported;
   }
 
+  Future<AvailableMobilePayOptions> availableMobilePayOptions({
+    IsGooglePaySupportedParams? googlePay,
+    PlatformPayWebPaymentRequestCreateOptions? webPaymentRequestCreateOptions,
+  }) async {
+    await _awaitForSettings();
+
+    return await _platform.availableMobilePayOptions(
+      params: googlePay,
+      paymentRequestOptions: webPaymentRequestCreateOptions,
+    );
+  }
+
   /// Launches the relevant native wallsheet (Apple Pay on iOS and Google Pay on Android)
   /// in order to create a payment intent
   ///
@@ -283,6 +295,16 @@ class Stripe {
       throw StripeError(message: error.message, code: error.message);
     }
   }
+
+  Future<PaymentMethod> createPaymentMethodWithElements() =>
+      _platform.createPaymentMethodWithElements();
+
+  Future<void> elementsSubmit() => _platform.elementsSubmit();
+
+  Future<PaymentIntent> handleCardAction(
+    final String paymentIntentClientSecret,
+  ) =>
+      _platform.handleCardAction(paymentIntentClientSecret);
 
   /// Creates a single-use token that represents a credit card’s details.
   ///
@@ -701,6 +723,7 @@ class Stripe {
   }
 
   bool _needsSettings = true;
+
   void markNeedsSettings() {
     _needsSettings = true;
     if (!_platform.updateSettingsLazily) {